Further updates to the Foundation Standard
